+++ app/policies/backoffice/service_policy.rb
@@ -27,6 +27,10 @@ class Backoffice::ServicePolicy < ApplicationPolicy
     owner?
   end
 
+  def published?
+    owner? && !record.status == "draft"
+  end
+
   def destroy?
     owner? && project_items.count.zero?
   end
@@ -41,7 +45,8 @@ class Backoffice::ServicePolicy < ApplicationPolicy
      :tutorial_url, :restrictions, :phase,
      :activate_message, :logo,
      [contact_emails: []], [research_area_ids: []],
-     [platform_ids: []], :tag_list, [category_ids: []]]
+     [platform_ids: []], :tag_list, [category_ids: []],
+     :status]
   end
 
   private